Version 0.1
The initial code for the game is correct, however currently the second "if" statement is causing an error where both are printed even though the second boolean statement has not been reached. Looking to resolve this issue in the next update, also looking to add a bit of user validation and if the user gets the number wrong, then the game will advise the user on the correct number.