Update 0.3
Updated the while loop to add in a "true" boolean variable, Inteliji however doesnt like the format of the method, unable to work out why these methods are creating error messages, will be looking to resolve this issue in the next update. furthermore the "while" loop doesnt want to be executed, attempting to discover why.